+++ SleepHunter.Updater/App.xaml.cs
@@ -1,13 +1,14 @@
 ﻿using System;
 using System.Diagnostics;
 using System.IO;
+using System.Threading.Tasks;
 using System.Windows;
 
 namespace SleepHunter.Updater
 {
     public partial class App : Application
     {
-        protected override void OnStartup(StartupEventArgs e)
+        protected override async void OnStartup(StartupEventArgs e)
         {
             // Invalid number of arguments, exit
             // Usage: Updater.exe <zip file> <install path>
@@ -19,34 +20,54 @@ namespace SleepHunter.Updater
 
             var updateFilePath = e.Args[0];
             var installationPath = e.Args[1];
-            var executableFile = Path.Combine(installationPath, "SleepHunter.exe");
 
             base.OnStartup(e);
 
             var mainWindow = new MainWindow();
+            mainWindow.RetryRequested += async (sender, _) =>
+            {
+                mainWindow.ResetState();
+
+                var retrySuccess = await PerformUpdate(mainWindow, updateFilePath, installationPath);
+                if (!retrySuccess)
+                    mainWindow.ToggleRetryButton(true);
+                else
+                    Shutdown();
+            };
+
             mainWindow.Show();
+
+            var success = await PerformUpdate(mainWindow, updateFilePath, installationPath);
+            if (!success)
+                mainWindow.ToggleRetryButton(true);
+            else
+                Shutdown();
+        }
+
+        private async Task<bool> PerformUpdate(MainWindow mainWindow, string updateFilePath, string installationPath)
+        {
+            var executableFile = Path.Combine(installationPath, "SleepHunter.exe");
 
             // Check that the update file exists, show error if missing
             if (!File.Exists(updateFilePath))
             {
                 mainWindow.SetStatusText("Unable to Update");
                 mainWindow.SetErrorMessage("Missing update file.\nYou can try again within SleepHunter, or install it manually.");
-                return;
+                return false;
             }
 
             // Terminate any existing SleepHunter instances
             try
             {
-                mainWindow.SetStatusText("Preparing...");
-                
-                foreach (var process in Process.GetProcessesByName("SleepHunter"))
-                    process.Kill();
+                mainWindow.SetStatusText("Waiting for SleepHunter...");
+                await Task.Delay(3000);
+                TerminateAllAndWait("SleepHunter");
             }
             catch (Exception ex)
             {
                 mainWindow.SetStatusText("Unable to Update");
                 mainWindow.SetErrorMessage(ex.Message);
-                return;
+                return false;
             }
 
             // Try to update, and display an error if something goes wrong
@@ -58,7 +79,7 @@ namespace SleepHunter.Updater
             {
                 mainWindow.SetStatusText("Unable to Update");
                 mainWindow.SetErrorMessage(ex.Message);
-                return;
+                return false;
             }
 
             // Check that the executable exists, show error if missing
@@ -66,14 +87,36 @@ namespace SleepHunter.Updater
             {
                 mainWindow.SetStatusText("Unable to Restart");
                 mainWindow.SetErrorMessage("Missing SleepHunter executable in installation folder.\nYou may need to apply the update manually.");
-                return;
+                return false;
             }
 
             // Restart SleepHunter
             mainWindow.SetStatusText("Restarting SleepHunter...");
-            Process.Start(executableFile);
+            try
+            {
+                Process.Start(executableFile);
+                return true;
+            }
+            catch
+            {
+                mainWindow.SetStatusText("Unable to Restart");
+                mainWindow.SetErrorMessage("Unable to restart SleepHunter automatically.\nYou can try launching it manually.");
+                return false;
+            }
+        }
 
-            Shutdown();
+        private async void TerminateAllAndWait(string processName)
+        {
+            var matchingProcesses = Process.GetProcessesByName(processName);
+
+            foreach (var process in matchingProcesses)
+                process.Kill();
+
+            while (matchingProcesses.Length > 0)
+            {
+                await Task.Delay(3000);
+                matchingProcesses = Process.GetProcessesByName(processName);
+            }
         }
     }
 }